requirements.txt
Flask>=1.1.2
pyserial>=3.4
Flask-APScheduler>=1.11.0